requested changes

+++ windows/pb_sleep_ms.c
@@ -1,7 +1,7 @@
 /* -*- c-file-style:"stroustrup"; indent-tabs-mode: nil -*- */
-#include "core/pb_sleep_ms.h"
+#include <windows.h>
 
-void pb_sleep_ms(DWORD dwMilliseconds)
+void pb_sleep_ms(unsigned long dwMilliseconds)
 {
     Sleep(dwMilliseconds);
 }
